Refactor to be more similar to other MCPs
