+++ src/allroadstoliterature/main.py
@@ -2,40 +2,28 @@ import pprint
 from typing import Dict, Any, Optional
 
 import click
-import habanero
+from fastmcp import FastMCP
+
+from src.allroadstoliterature.tools import get_doi_metadata
 
 
-def get_doi_metadata(doi: str) -> Optional[Dict[str, Any]]:
-    """
-    Retrieve metadata for a scientific article using its DOI.
+def create_mcp():
+    """Create the FastMCP instance with registered tools."""
+    mcp = FastMCP("all-roads-to-literature")
 
-    Args:
-        doi: The Digital Object Identifier of the article.
+    # Register all tools
+    mcp.tool(get_doi_metadata)
 
-    Returns:
-        A dictionary containing the article metadata if successful, None otherwise.
-    """
-    cr = habanero.Crossref()
-    try:
-        result = cr.works(ids=doi)
-        return result
-    except Exception as e:
-        print(f"Error retrieving metadata for DOI {doi}: {e}")
-        return None
+    return mcp
 
 
-@click.command()
-@click.option('--doi', '-d', required=True, help='DOI of the article to retrieve metadata for')
-def main(doi: str) -> None:
-    """Retrieve and display metadata for a scientific article using its DOI."""
-    click.echo(f"Retrieving metadata for DOI: {doi}")
+# Create the FastMCP instance at module level
+mcp = create_mcp()
 
-    metadata = get_doi_metadata(doi)
 
-    if metadata:
-        pprint.pprint(metadata)
-    else:
-        click.echo("Failed to retrieve metadata.")
+def main():
+    """Main entry point for the application."""
+    mcp.run()
 
 
 if __name__ == "__main__":
